- Database Delete Count column 	- Authorization Delete Taxonomy

+++ Http/Controllers/TaxonomyController.php
@@ -242,6 +242,7 @@ class TaxonomyController extends CoreController
     public function destroy(Request $request)
     {
         $query = $this->taxonomy_m->findOrFail(decrypt($request->input('id')));
+        $this->authorize('delete-taxonomy', $query);
 
         try {
             if($query->delete())
